fixes from pull request
