+++ src/app/sso-api.service.ts
@@ -1,7 +1,8 @@
-import { Injectable } from '@angular/core';
-import { LoginFormModel, User } from './shared/model'
-import { Observable } from 'rxjs/Rx';
+import { Injectable } from '@angular/core'
+import { User } from './shared/model'
+import { Observable } from 'rxjs/Rx'
 import { Http } from '@angular/http'
+import { LoginFormModel } from './login/login.model'
 
 const userEmail: string = 'user@mail.ua'
 const userPassword: string = 'Pass1234'
@@ -16,8 +17,7 @@ export class SSOApiService {
         .map(res => res.json() || null)
         .catch(err => Observable.throw(err))
     }
-    //Не знаю как создать фейковый Observable, делаем почти реальный запрос
-    return this.http.get('assets/nouser').map(res => null)
+    return Observable.of(undefined)
   }
