--- src/assets/nouser
@@ -1,1 +1,0 @@
-{}
